requirements.txt
langchain
spotipy
python-dotenv
openai
agentops